update - becky’s changes, also no more circular dependencies
